Typecheck rm

+++ papis/commands/rm.py
@@ -8,7 +8,6 @@ Cli
 import papis
 import papis.pick
 import papis.utils
-import papis.config
 import papis.document
 import papis.cli
 import papis.strings
@@ -18,30 +17,33 @@ import click
 import logging
 import os
 
+from typing import Optional
 
-def run(document, filepath=None, git=False):
+
+def run(document: papis.document.Document,
+        filepath: Optional[str] = None,
+        git: bool = False) -> None:
     """Main method to the rm command
     """
     db = papis.database.get()
+    _doc_folder = document.get_main_folder()
+    if not _doc_folder:
+        raise Exception(papis.strings.no_folder_attached_to_document)
     if filepath is not None:
         os.remove(filepath)
         document['files'].remove(os.path.basename(filepath))
         document.save()
         db.update(document)
         if git:
-            papis.git.rm(document.get_main_folder(), filepath)
-            papis.git.add(document.get_main_folder(), document.get_info_file())
-            papis.git.commit(
-                document.get_main_folder(),
-                "Remove file '{0}'".format(filepath))
+            papis.git.rm(_doc_folder, filepath)
+            papis.git.add(_doc_folder, document.get_info_file())
+            papis.git.commit(_doc_folder, "Remove file '{0}'".format(filepath))
     else:
         if git:
-            _topfolder = os.path.dirname(
-                os.path.abspath(document.get_main_folder()))
-            papis.git.rm(
-                document.get_main_folder(), document.get_main_folder(),
-                recursive=True)
-            papis.git.commit(_topfolder,
+            _topfolder = os.path.dirname(os.path.abspath(_doc_folder))
+            papis.git.rm(_doc_folder, _doc_folder, recursive=True)
+            papis.git.commit(
+                _topfolder,
                 "Remove document '{0}'".format(
                     papis.document.describe(document)))
         else:
@@ -65,7 +67,13 @@ def run(document, filepath=None, git=False):
     is_flag=True,
     default=False)
 @papis.cli.all_option()
-def cli(query, git, _file, force, _all, sort_field, sort_reverse):
+def cli(query: str,
+        git: bool,
+        _file: bool,
+        force: bool,
+        _all: bool,
+        sort_field: Optional[str],
+        sort_reverse: bool) -> None:
     """Delete a document or a file"""
 
     documents = papis.database.get().query(query)
@@ -77,7 +85,7 @@ def cli(query, git, _file, force, _all, sort_field, sort_reverse):
 
     if not documents:
         logger.warning(papis.strings.no_documents_retrieved_message)
-        return 0
+        return
 
     if not _all:
         documents = [papis.pick.pick_doc(documents)]
@@ -90,7 +98,8 @@ def cli(query, git, _file, force, _all, sort_field, sort_reverse):
                 continue
             if not force:
                 tbar = 'The file {0} would be removed'.format(filepath)
-                if not papis.utils.confirm("Are you sure?", bottom_toolbar=tbar):
+                if not papis.utils.confirm(
+                        "Are you sure?", bottom_toolbar=tbar):
                     continue
             logger.info("Removing %s..." % filepath)
             run(document, filepath=filepath, git=git)
@@ -104,7 +113,8 @@ def cli(query, git, _file, force, _all, sort_field, sort_reverse):
                     title=tbar,
                     text=papis.document.dump(document),
                     lexer_name='yaml')
-                if not papis.utils.confirm("Are you sure?", bottom_toolbar=tbar):
+                if not papis.utils.confirm(
+                        "Are you sure?", bottom_toolbar=tbar):
                     continue
             logger.warning("removing ...")
             run(document, git=git)
